Transfer some GUI formatting to ui package
