+++ build.gradle
@@ -47,7 +47,7 @@ application {
 }
 
 shadowJar {
-    archiveBaseName = "launcher"
+    archiveBaseName = "grok"
     archiveClassifier = null
 }
